Correct the order of the arguments for a String(format:) call
Official printf spec (which NSString uses) specifies the precision field must appear before the argument.
The call worked even though the arguments were not correctly ordered because NSString (and maybe printf?) probably tries to be clever and uses the Double value for the Double argument and the Int for the Int argument. Or maybe it’s the compiler… (I think that’s the compiler that does the Float to Double conversion already). Anyway, the order was definitely incorrect!

FYI, the printf specs: https://pubs.opengroup.org/onlinepubs/009695399/functions/printf.html